--- src/rest/RestResults.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const RestResults = (props) => {
-    return (
-        <div>
-            {props.results.map(result => {
-                
-            })}
-        </div>
-    )
-}
